refactor: перенести класс ErrorResponse в ExceptionHandler
