--- src/main/java/ru/yandex/practicum/filmorate/model/ErrorResponse.java
@@ -1,13 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-public class ErrorResponse {
-    private final String error;
-
-    public ErrorResponse(String error) {
-        this.error = error;
-    }
-
-    public String getError() {
-        return error;
-    }
-}
